Related work items: #242, #245, #246, #247, #248

+++ RookieOnlineAssetManagement/Controllers/AssetController.cs
@@ -31,19 +31,21 @@ namespace RookieOnlineAssetManagement.Controllers
             return Ok(result.Datas);
         }
         [HttpGet("assignment-asset")]
-        public async Task<ActionResult<IEnumerable<AssetModel>>> GetListForAssignmentAsync(string currentAssetId, string query, SortBy? AssetIdSort, SortBy? AssetNameSort, SortBy? CategoryNameSort)
+        public async Task<ActionResult<IEnumerable<AssetModel>>> GetListForAssignmentAsync([FromQuery] AssetAssignmentRequestParams requestParams)
         {
-            var locationId = RequestHelper.GetLocationSession(HttpContext);
-            return Ok(await _assetService.GetListAssetForAssignmentAsync(currentAssetId, locationId, query, AssetIdSort, AssetNameSort, CategoryNameSort));
+            requestParams.LocationId = RequestHelper.GetLocationSession(HttpContext);
+            return Ok(await _assetService.GetListAssetForAssignmentAsync(requestParams));
         }
         [HttpGet("history")]
         public async Task<ActionResult<IEnumerable<AssetHistoryModel>>> GetListHistoryAsync(string assetId)
         {
+            if (string.IsNullOrEmpty(assetId)) return BadRequest("Asset Id is not valid");
             return Ok(await _assetService.GetListAssetHistoryAsync(assetId));
         }
         [HttpGet("{id}")]
         public async Task<ActionResult<AssetDetailModel>> GetAsync(string id)
         {
+            if (string.IsNullOrEmpty(id)) return BadRequest("Asset Id is not valid");
             return Ok(await _assetService.GetAssetByIdAsync(id));
         }
         [HttpPost]
@@ -57,12 +59,14 @@ namespace RookieOnlineAssetManagement.Controllers
         public async Task<ActionResult<AssetModel>> UpdateAsync(string id, AssetRequestModel assetRequestModel)
         {
             assetRequestModel.LocationId = RequestHelper.GetLocationSession(HttpContext);
+            if (string.IsNullOrEmpty(id)) return BadRequest("Asset Id is not valid");
             if (!ModelState.IsValid) return BadRequest(assetRequestModel);
             return Ok(await _assetService.UpdateAssetAsync(id, assetRequestModel));
         }
         [HttpDelete("{id}")]
         public async Task<ActionResult<bool>> DeleteAsync(string id)
         {
+            if (string.IsNullOrEmpty(id)) return BadRequest("Asset Id is not valid");
             return Ok(await _assetService.DeleteAssetAsync(id));
         }
     }
